+++ src/etphoneshome/managers/EntityManager.java
@@ -68,19 +68,19 @@ public class EntityManager {
      *
      * @param amount The amount of {@code Enemy} to be spawned
      */
-    public void spawnRandomEntities(int amount) {
+    public void spawnRandomEntities(double amount) {
         Random random = new Random();
-        int xCord = 1728/2 + 70;
-        for (int i = 0; i < amount; i++) {
-            xCord = (random.nextInt(1920) + (int) (new Police()).getLeftEntitySprite().getWidth())+xCord;
-            int type = random.nextInt(2);
+        double xCord = 1728/2 + 70;
+        for (double i = 0; i < amount; i++) {
+            xCord = (random.nextInt(1920) + (new Police()).getLeftEntitySprite().getWidth())+xCord;
+            double type = random.nextInt(2);
             Enemy enemy = new Police();
             if (type == 0) {
                 enemy = new Police();
             } else {
                 enemy = new Scientist();
             }
-            enemy.setLocation(new Location(xCord, UILauncher.getGraphicsRepainter().HEIGHT - 100 - (int) (new Police()).getLeftEntitySprite().getHeight()));
+            enemy.setLocation(new Location(xCord, UILauncher.getGraphicsRepainter().HEIGHT - 100 - (new Police()).getLeftEntitySprite().getHeight()));
             this.addEnemy(enemy);
         }
     }
