+++ src/etphoneshome/objects/Velocity.java
@@ -7,12 +7,12 @@ public class Velocity {
     /**
      * The amount of horizontal velocity. A negative or positive will indicate direction
      */
-    private int horizontalVelocity = 0;  //Velocity is a vector so negative velocity is possible
+    private double horizontalVelocity = 0;  //Velocity is a vector so negative velocity is possible
 
     /**
      * The amount of vertical velocity. A negative or positive will indicate direction
      */
-    private int verticalVelocity = 0;
+    private double verticalVelocity = 0;
 
     /**
      * Simple default constructor which will use the default horizontal/vertical velocity values of 0
@@ -28,7 +28,7 @@ public class Velocity {
      * @param horizontalVelocity The amount of horizontal velocity
      * @param verticalVelocity   The amount of vertical velocity
      */
-    public Velocity(int horizontalVelocity, int verticalVelocity) {
+    public Velocity(double horizontalVelocity, double verticalVelocity) {
         this.horizontalVelocity = horizontalVelocity;
         this.verticalVelocity = verticalVelocity;
     }
@@ -38,7 +38,7 @@ public class Velocity {
      *
      * @return the amount of horizontal velocity
      */
-    public int getHorizontalVelocity() {
+    public double getHorizontalVelocity() {
         return this.horizontalVelocity;
     }
 
@@ -47,7 +47,7 @@ public class Velocity {
      *
      * @return the amount of vertical velocity
      */
-    public int getVerticalVelocity() {
+    public double getVerticalVelocity() {
         return this.verticalVelocity;
     }
 
@@ -56,7 +56,7 @@ public class Velocity {
      *
      * @param amount the amount of horizontal velocity
      */
-    public void setHorizontalVelocity(int amount) {
+    public void setHorizontalVelocity(double amount) {
         this.horizontalVelocity = amount;
     }
 
@@ -65,7 +65,7 @@ public class Velocity {
      *
      * @param amount the amount of vertical velocity
      */
-    public void setVerticalVelocity(int amount) {
+    public void setVerticalVelocity(double amount) {
         this.verticalVelocity = amount;
     }
 
@@ -74,7 +74,7 @@ public class Velocity {
      *
      * @param amount The amount the horizontal velocity changes and in which direction
      */
-    public void changeHorizontalVelocity(int amount) {
+    public void changeHorizontalVelocity(double amount) {
         this.horizontalVelocity += amount;
     }
 
@@ -83,7 +83,7 @@ public class Velocity {
      *
      * @param amount The amount the vertical velocity changes and in which direction
      */
-    public void changeVerticalVelocity(int amount) {
+    public void changeVerticalVelocity(double amount) {
         this.verticalVelocity += amount;
     }
